refactor(utils): move usePrefersReducedMotion out of CyclingText
